+++ src/theme/components/notification.ts
@@ -1,7 +1,7 @@
 import { NotificationTypes } from "../../components/Notification/NotificationTypes";
 
 const Notification = {
-  parts: ["container", "dismissibleButton", "dismissibleIcon", "icon"],
+  parts: ["container", "dismissibleButton", "icon"],
   baseStyle: ({ centered, noMargin, notificationType }) => {
     let bg = "ui.status.primary";
     if (notificationType === NotificationTypes.Announcement) {
@@ -24,24 +24,23 @@ const Notification = {
         maxWidth: "var(--breakpoint-large)",
       },
       dismissibleButton: {
+        border: "none",
         alignItems: "center",
         bgColor: "inherit",
         color: "ui.black",
         display: "flex",
-        height: { base: "var(--space-l)", lg: "var(--space-m)" },
-        width: { base: "var(--space-l)", lg: "var(--space-m)" },
-        padding: "0",
+        height: "32px",
+        width: "32px",
+        minWidth: "0",
         position: "absolute",
-        right: "xs",
-        top: "xs",
+        right: "0",
+        top: "0",
+        svg: {
+          marginTop: "0",
+        },
         _hover: {
           bg: "inherit",
         },
-      },
-      dismissibleIcon: {
-        height: "var(--space-m)",
-        marginRight: "0",
-        width: "var(--space-m)",
       },
       icon: {
         flexShrink: "0",